% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_bar_intr.R
\name{ps_bar_intr}
\alias{ps_bar_intr}
\title{Title}
\usage{
ps_bar_intr(ps, t_lvl, n)
}
\arguments{
\item{ps}{A phyloseq object populated with an OTU table, an assigned taxa table, and sample data (with a column named "sample.name")}

\item{t_lvl}{The desired taxonomic level (e.g. Phylum, Class, Order, Genus)}

\item{n}{The number of most relatively abundant OTUs to plot (recommended start ~20)}
}
\value{
An interactive stacked bar plot
}
\description{
Title
}
\examples{
z <- cbind(c("Archaea","Methyltropha"),c("Bacteria","Bacteriodes"))
colnames(z) <- c("Phylum", "Class")
df <- data.frame(c("A", "B"), c("C", "D"))
colnames(df) <- c("sample.name", "group")
y <- matrix(1:4,nrow=2)
ps <- phyloseq::phyloseq(phyloseq::otu_table(y, taxa_are_rows = FALSE),
phyloseq::tax_table(z),phyloseq::sample_data(df))

t_lvl <- "Class"
n <- 20
ps_bar_intr(ps, t_lvl, n)
}
